Established "Static" Directory
Instead of conflating static files with the development build, separated
static files into their own directory. This reduces ambiguity over what
files are generated from a compilation process (and, thus, may need to
be cleaned between builds) and which are part of the development
workflow. It also allows the /Builds directory to be excluded from
source control (e.g., via a .gitignore file).